#34 fix: Refactor SpeechmaticsMessages structure
New structure is now:
- `Data.SpeechmaticsMessages` is a namespace of namespaces, no classes at its level
  - It made sense to me to not dump all of the messages straight into `Data`
  - If the message classes are directly in `Data.SpeechmaticsMessages`, then we can't make a namespace with the same name
- All messages live in their own `Data.SpeechmaticsMessages.<message type>` namespace
- If a message class has sub-classes, then the sub-class lives in `<parent namespace>.<JSON attribute name>`

Also includes a rename of `StartRecognitionMessage_AudioType` + associated variables to `StartRecognitionMessage_AudioFormat`.
The object is being referred to in the documentation as `AudioType`, but it's being used for the `audio_format` attribute.
For consistency with all other subclasses, the names have been adjusted to match the attribute its used for.